+++ CHANGELOG.md
@@ -1,9 +1,18 @@
+0.4.2 (2022-01-07)
+==================
+
+Fixed
+-----
+- Fix the HTML output when `eval_functions=True`
+- Display "Loading..." under the table header until the table is displayed with datatables.net
+- `init_notebook_mode(all_interactive=False)` restores the original Pandas HTML representation.
+
 0.4.1 (2022-01-06)
 ==================
 
 Fixed
 -------
-- Long column names don't overlap any more (#28)
+- Long column names don't overlap anymore (#28)
 
 
 0.4.0 (2022-01-06)